new type of parsing

+++ pushSwap/inc/push_swap.h
@@ -6,7 +6,7 @@
 /*   By: kgebski <kgebski@student.42wolfsburg.de    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/30 11:50:52 by kgebski           #+#    #+#             */
-/*   Updated: 2023/05/31 14:25:21 by kgebski          ###   ########.fr       */
+/*   Updated: 2023/05/31 18:03:55 by kgebski          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@
 int		*parser(int argc, char *argv[]);
 int		not_unique(int *tab, size_t size);
 void	initialize_list(int *tab, size_t size, t_list **lst);
-void	sort_stack(t_list **stack_a, t_list **stack_b, size_t size);
+void	sort_stack(t_list **stack_a, t_list **stack_b, int size);
 
 int		is_sorted(t_list *lst);
 void	sort_three(t_list **stack_a);
